Instance Language Configuration
I have provided a `lang` method on both moment.fn and moment.duration.fn
to allow for instances of moments and durations to have their own
language configuration.  Additionally, the `lang` method is a getter
which returns either the moment's language definition object or the
global language definition object if the moment did not have one set.

Also, I modified the code surrounding meridiem to always expect a
function.  This simplifies the formatter code, and also ensures
inheritability of language configuration values.