getLangDefinition will now automatically load lang
I extracted the loadLang function out of moment.lang so that it could be
used internally in moment.  When developers use moment.fn.lang or
moment.langData now, if the language requested has not yet been loaded,
it will be loaded.  Unlike moment.lang, these functions do not change
the global language config.