Diffs now return durations.
As discussed in #282, the diff function now returns a duration instead
of just a millisecond number for the one parameter incantation.
Internally, diff uses the duration to perform the math.  This commit
passes all but two of the unit tests (fails regarding returning years as
floats).  This is not ready to be merged in.