Fix parsing/formatting of first century dates.
According to ECMA-262, `new Date(y, …)` will auto-convert `y` to `1900 +
y` if `0 ≤ y ≤ 99`, hence setFullYear or setUTCFullYear is used instead.

This also zero-pads the year for the "YYYY" format, since the year may
not necessarily have four digits.

Lastly, the pt-br test is adjusted to test for the first day of the
month instead of the zeroth (which was causing it to fail).