Removed a failing iso-format parsing test
One of the things that is broken on Android is "including a trailing UTC in the
input should work", which is a test added a long time ago, that used the fact
that ISO format matched beginning of string.

So right now this test goes through Date, which is different on regular vs
mobile browser. Also adding UTC at the end doesn't make it parse a UTC date, so
its misleading to say the least.

I think removing the test makes the most sense in this case. The behavior (of
adding UTC at the end is not documented, so shouldn't break legitimate code).