Changed Swedish locale.
Changed the Swedish locale for lastWeek to something more like what one
would say. As it was before with for instance "Förra söndagen" (Last
Sunday) while referring to the closest passed Sunday is not what is used
in language, last Sunday refers to the Sunday before the closest passed
one, so I changed it to that.