Update nb.js time format for LT and LTS
The time format for LT and LTS do not align with what is the default time settings in Windows for Norwegian Bokmål and what most people use. Propose H:mm instead of H.mm and LT:ss instead of LT.ss. Most norwegians do not use a 13.37 time format, but rather 13:37.