Have found several references which cites the official abbreviation statements
http://i18n.skolelinux.no/localeoppsett.html is the resource most norwegian
software localizers use.

For months, we have the full version and the short version which does not
abbreviate the shortest month names. Abbreviations ends with a ".".

For weekdays, we have the full version and the abbreviated version. All
abbreviations should be two letters and a ".". I let the minimal be the
abbreviated version without the ".", and assume this will be used where it is
obvious that it is an abbreviation.