Fix test broken only in March
Even if you set custom rounding threshold for days -> month, because relative
time works on durations, which computes it from two moments, if its over
a month in real terms (1st Feb to 2nd March) it will be shown as a month no
matter the thres. It might be considered a bug, but for now it's a feature.