[bugfix] Fix warnings flagged by closure compiler (fixes #4347)
This statement as it stands:

  this.weekdays(mom, '').replace('.', '\.?')

is not correct, and recent closure compiler issues unnecessary escape
warning, claiming that '\.' is equivalent to just '.'.

It was actually intended the RegExp to search for a literal period. In
that case the correct fix would be to double the backslashes rather than
removing them.