Update typing for diff function
This might be the intended typing, but it's not accurate. No error is thrown by passing anything other than a `MomentInput` and the function still returns a `number` type as `NaN` is categorised as a `number`. Either the implementation needs to change such that this throws without a `MomentInput` or this typing needs to change, as `Nan` is a valid return value.