Update mr.js
We Marathi people follow standards mentioned below, to label timespans.

| पहाटे (post midnight, pre-sunrise) |    >= 12am to < 6am |
| सकाळी (post sunrise, pre-noon) |    >= 6am to < 12pm |
| दुपारी (afternoon) |    >= 12pm to < 5pm |
| सायंकाळी (evening) |    >= 5pm to < 8pm |
| रात्री (night) |    >= 8pm to < 12am |

Actually, 12am is confusing in Marathi tradition. In modern clock system, Moment 12am is considered in arriving date as 0am. And in Marathi tradition it is referred to as final moment of the previous date (Sort of, 0:01 is starting minute of the arriving date) . To comply with modern clock system, I propose to refer to 0am as पहाटे , so that it is considered as moment in the arriving date. Hope this discription itself is not confusing :)