[locale] eo: Fixes tests to match new expressions
Adding the letter "t" for the months "marto" and "septembro" to distinguish them from "mardo" (=tuesday) and "sep" (=seven)
Correction of some expressions in which accusative is needed.
Adding the word "kelkaj" (=some), because "post sekundoj" and "antaŭ sekundoj" don't really make sense.
Adding the word "unu" (=one) because it's about the quantity "1", and not about an indefinite thing introduced by the article a/an.