[locale] sk: Improve future tense (#5206)
"Za" is commonly (and _incorrectly_) used to relatively refer to a moment in the future. This error is mainly influenced by Czech language, which is very similar and most Slovaks will understand it, many times domesticating phrases from Czech language. In Czech, "za" is used, though in Slovak, "o" must be used.

Here are references to both "za" and "o" in the official dictionary of Slovak language:
https://slovnik.juls.savba.sk/?w=o&s=exact&d=kssj4
https://slovnik.juls.savba.sk/?w=za&s=exact&d=kssj4
And here is a more readable unofficial explanation: https://www.pravopisne.sk/o-hodinu-x-za-hodinu/