[tests] Add test(s) for the weird days of JAN2022
In the beginning of 2022, there are a few “weird days” where the week numbers in various methods don’t line up because `Moment.js` is internally working with ISO-8601 week numbers whereas the default locale is `EN_US`. This is intentional behaviour.

This discrepancy between definitions means that developers using `Moment.js` are possibly unaware of the implementation differences. This commit adds a test which is expected to confirm developers’ knowledge regarding the API contract on this particular instance.

ISO-8601 defines that
- the week starts on monday and
- the first week of a year is the one with the first thursday

The `EN_US`-locale defines that
- the week starts on sunday and
- the first week of a year is the one with “January 1st”